Add a nojekyll file
Images are not rendering properly in the rendered bookdown  document
hosted on github pages, but the files exist. Apparently the jekyll
processing on GitHub pages ignores any files or directories starting
with an underscore. However, for this bookdown, all the images are
hosted in a directory called "_main_files". Because it starts with
an underscore, it is ignored. Try adding a .nojekyll file as
recommended by a couple sites:

https://github.blog/2009-12-29-bypassing-jekyll-on-github-pages/
https://community.rstudio.com/t/broken-image-links-when-publishing-bookdown-on-github-pages/90897